# backend/src/main/java/com/hackathon/leave/controller/HolidayController.java
package com.hackathon.leave.controller;

import com.hackathon.leave.dto.HolidayDto;
import com.hackathon.leave.dto.HolidayRequest;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import java.util.List;
import org.springframework.http.ResponseEntity;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.web.bind.annotation.*;

@Tag(name = "Holidays")
@SecurityRequirement(name = "bearerAuth")
@RestController
@RequestMapping("/api/holidays")
public class HolidayController {

    @Operation(summary = "All holidays, by date")
    @PreAuthorize("isAuthenticated()")
    @GetMapping
    public ResponseEntity<List<HolidayDto>> list() {
        return Stubs.notImplemented();
    }

    @Operation(summary = "Add a holiday (HR only)")
    @PreAuthorize("hasRole('HR')")
    @PostMapping
    public ResponseEntity<HolidayDto> create(@Valid @RequestBody HolidayRequest request) {
        return Stubs.notImplemented();
    }
}